@@ -13,12 +13,13 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "type:check": "tsc --pretty --noEmit"
+    "type:check": "tsc --pretty --noEmit",
+    "lint": "eslint . --ext ts --ext tsx --fix"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0"
   },
   "devDependencies": {
